Completed game logic to turn off game board clickability when a game ends.
Added lines to remove the '.clickable' class from each gameboard box when a game is completed in a win, draw, or restart. Previously, squares were clickable after a game ended, which would trigger an attempt to claim the space clicked on. Server was properly updated, but game UI would still update. It was found that the '.clickable' class could only be successfully removed when done so from within the game/events.js file, but would not work when called from game/ui.js.